requirements.txt
aiofiles==24.1.0
fastapi
gdown==5.2.0
httpx==0.28.1
pydantic
python-dotenv==1.1.0
rich==14.0.0
uvicorn